+++ src/main/java/me/desht/pneumaticcraft/common/thirdparty/jei/JEISpecialCraftingCategory.java
@@ -11,9 +11,7 @@ import net.minecraft.client.resources.I18n;
 import java.util.List;
 
 public abstract class JEISpecialCraftingCategory<T extends IRecipeWrapper> extends PneumaticCraftCategory<T> {
-
     private List<String> text;
-    private final List<MultipleInputOutputRecipeWrapper> allRecipes = getAllRecipes();
 
     JEISpecialCraftingCategory(IJeiHelpers jeiHelpers) {
         super(jeiHelpers);
@@ -37,7 +35,7 @@ public abstract class JEISpecialCraftingCategory<T extends IRecipeWrapper> exten
     public void drawExtras(Minecraft minecraft) {
         if (text != null) {
             for (int i = 0; i < text.size(); i++) {
-                Minecraft.getMinecraft().fontRenderer.drawString(text.get(i), 5, 24 + i * 10, 0xFF000000);
+                Minecraft.getMinecraft().fontRenderer.drawString(text.get(i), 5, 24 + i * 10, 0xFF404040);
             }
         }